+++ nmtpy/models/fusion_concat_dep_ind.py
@@ -565,6 +565,6 @@ class Model(ParentModel):
         self.f_next = theano.function(inputs, outs, name='f_next')
 
     def get_alpha_regularizer(self, alpha_c):
-        alpha_c = theano.shared(np.float32(alpha_c), name='alpha_c')
+        alpha_c = theano.shared(np.float64(alpha_c).astype(FLOAT), name='alpha_c')
         alpha_reg = alpha_c * ((1.-self.alphas[1].sum(0))**2).sum(0).mean()
         return alpha_reg
